fix: do not try to connect to a DB if password/username not available (not on a DIRAC server)
