+++ src/DIRAC/WorkloadManagementSystem/Client/ServerUtils.py
@@ -1,28 +1,9 @@
 """
-  Provide uniform interface to backend for local and remote clients.return
-
-  There's a pretty big assumption here: that DB and Handler expose the same calls, with identical signatures.
-  This is not exactly the case for WMS DBs and services.
+This module is used to create an appropriate object which can be used to insert records to the WMS.
+It always try to insert the records directly. In case of failure a WMS client is used...
 """
 
-
-def getDBOrClient(DB, serverName):
-    """Tries to instantiate the DB object
-    and returns it if we manage to connect to the DB,
-    otherwise returns a Client of the server
-    """
-    from DIRAC import gLogger
-    from DIRAC.Core.Base.Client import Client
-
-    try:
-        myDB = DB()
-        if myDB._connected:
-            return myDB
-    except Exception:
-        pass
-
-    gLogger.info("Can not connect to DB will use %s" % serverName)
-    return Client(url=serverName)
+from DIRAC.Core.Utilities.ServerUtils import getDBOrClient
 
 
 def getPilotAgentsDB():
